requirements.txt
selenium==4.19.0
prettytable==3.7.0
python-dotenv==1.0.0
requests==2.31.0
pytz==2024.1